Refactor: redefined line data creation
