+++ app/(components)/Search/Hero.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+export const Hero = () => {
+  return (
+    <>
+      <h1 className="text-center text-3xl text-neutral-200 mb-2 sm:text-6xl">
+        Routify
+      </h1>
+      <h2 className="mb-5 sm:mb-10 text-neutral-500 text-center pl-2 pr-2 text-xs sm:text-lg">
+        Your ultimate source for pathfinding algorithms in any city.
+      </h2>
+    </>
+  );
+};
