+++ lib/services/parsing.js
@@ -1,38 +1,15 @@
-import { latlonToXY } from "../utilities/coordUtils";
+// Extract the nodes and ways from the response
+export function parseOverpassResponse(jsonData) {
+  let nodes = [];
+  let ways = [];
 
-const parseLineData = (jsonData, boundingbox) => {
-  let returnArray = [];
-
-  let north = boundingbox[0];
-  let south = boundingbox[1];
-  let east = boundingbox[2];
-  let west = boundingbox[3];
-
-  let mapWidth = Math.abs(east - west) * 450;
-  let mapHeight = Math.abs(north - south) * 450;
-
-  jsonData["elements"].forEach((obj) => {
-    const currentList = obj["geometry"];
-
-    for (let i = 0; i < currentList.length - 1; i++) {
-      const start = latlonToXY(
-        currentList[i].lat,
-        currentList[i].lon,
-        mapWidth,
-        mapHeight
-      );
-      const end = latlonToXY(
-        currentList[i + 1].lat,
-        currentList[i + 1].lon,
-        mapWidth,
-        mapHeight
-      );
-
-      returnArray.push({ start, end });
+  jsonData["elements"].forEach((element) => {
+    if (element.type === "node") {
+      nodes.push({ id: element.id, lat: element.lat, lon: element.lon });
+    } else if (element.type === "way") {
+      ways.push({ nodes: element.nodes });
     }
   });
 
-  return returnArray;
-};
-
-export default parseLineData;
+  return { nodes, ways };
+}
