+++ lib/services/request.js
@@ -31,10 +31,10 @@ export default function sendRequest(
     const wayFilter = roadStrict;
     const timeout = 900;
     const maxHeapSize = 1073741824;
-    const responseType = "geom";
+    const responseType = "skel";
 
     // OverpassQL query to filter and extract roads from Overpass API
-    const queryString = `[timeout:${timeout}][maxsize:${maxHeapSize}][out:json];area(${areaId});(._; )->.area;(way${wayFilter}(area.area););out ${responseType};`;
+    const queryString = `[timeout:${timeout}][maxsize:${maxHeapSize}][out:json];area(${areaId});(._; )->.area;(way${wayFilter}(area.area); node(w));out ${responseType};`;
 
     const postData = {
       method: "POST",
