+++ lib/testing/test-cities/test-nomatium-results/test.js
@@ -1,0 +1,126 @@
+export const test1 = [
+  {
+    place_id: 325723095,
+    licence:
+      "Data © OpenStreetMap contributors, ODbL 1.0. http://osm.org/copyright",
+    osm_type: "relation",
+    osm_id: 182706,
+    lat: "39.9622601",
+    lon: "-83.0007065",
+    class: "boundary",
+    type: "administrative",
+    place_rank: 16,
+    importance: 0.6294399106820546,
+    addresstype: "city",
+    name: "Columbus",
+    display_name: "Columbus, Franklin County, Ohio, United States",
+    boundingbox: ["39.8086936", "40.1573082", "-83.2101797", "-82.7713119"],
+  },
+
+  {
+    place_id: 301899,
+    licence:
+      "Data © OpenStreetMap contributors, ODbL 1.0. http://osm.org/copyright",
+    osm_type: "relation",
+    osm_id: 1850813,
+    lat: "32.4610708",
+    lon: "-84.9880449",
+    class: "boundary",
+    type: "administrative",
+    place_rank: 16,
+    importance: 0.5407342831976765,
+    addresstype: "city",
+    name: "Columbus",
+    display_name: "Columbus, Muscogee County, Georgia, United States",
+    boundingbox: ["32.3738385", "32.6081231", "-85.0792070", "-84.6674630"],
+  },
+
+  {
+    place_id: 801525,
+    licence:
+      "Data © OpenStreetMap contributors, ODbL 1.0. http://osm.org/copyright",
+    osm_type: "relation",
+    osm_id: 2528689,
+    lat: "34.2814497",
+    lon: "-78.666593",
+    class: "boundary",
+    type: "administrative",
+    place_rank: 12,
+    importance: 0.523828114359475,
+    addresstype: "county",
+    name: "Columbus County",
+    display_name: "Columbus County, North Carolina, United States",
+    boundingbox: ["33.9444786", "34.4848477", "-79.0712116", "-78.1622008"],
+  },
+
+  {
+    place_id: 324535954,
+    licence:
+      "Data © OpenStreetMap contributors, ODbL 1.0. http://osm.org/copyright",
+    osm_type: "relation",
+    osm_id: 127528,
+    lat: "39.2014405",
+    lon: "-85.9213796",
+    class: "boundary",
+    type: "administrative",
+    place_rank: 16,
+    importance: 0.4947526395032379,
+    addresstype: "city",
+    name: "Columbus",
+    display_name: "Columbus, Bartholomew County, Indiana, United States",
+    boundingbox: ["39.1205960", "39.2748477", "-86.0029791", "-85.8508960"],
+  },
+  {
+    place_id: 324535954,
+    licence:
+      "Data © OpenStreetMap contributors, ODbL 1.0. http://osm.org/copyright",
+    osm_type: "relation",
+    osm_id: 127528,
+    lat: "39.2014405",
+    lon: "-85.9213796",
+    class: "boundary",
+    type: "administrative",
+    place_rank: 16,
+    importance: 0.4947526395032379,
+    addresstype: "city",
+    name: "Columbus",
+    display_name: "Columbus, Bartholomew County, Indiana, United States",
+    boundingbox: ["39.1205960", "39.2748477", "-86.0029791", "-85.8508960"],
+  },
+  {
+    place_id: 324535954,
+    licence:
+      "Data © OpenStreetMap contributors, ODbL 1.0. http://osm.org/copyright",
+    osm_type: "relation",
+    osm_id: 127528,
+    lat: "39.2014405",
+    lon: "-85.9213796",
+    class: "boundary",
+    type: "administrative",
+    place_rank: 16,
+    importance: 0.4947526395032379,
+    addresstype: "city",
+    name: "Columbus",
+    display_name: "Columbus, Bartholomew County, Indiana, United States",
+    boundingbox: ["39.1205960", "39.2748477", "-86.0029791", "-85.8508960"],
+  },
+  {
+    place_id: 324535954,
+    licence:
+      "Data © OpenStreetMap contributors, ODbL 1.0. http://osm.org/copyright",
+    osm_type: "relation",
+    osm_id: 127528,
+    lat: "39.2014405",
+    lon: "-85.9213796",
+    class: "boundary",
+    type: "administrative",
+    place_rank: 16,
+    importance: 0.4947526395032379,
+    addresstype: "city",
+    name: "Columbus",
+    display_name: "Columbus, Bartholomew County, Indiana, United States",
+    boundingbox: ["39.1205960", "39.2748477", "-86.0029791", "-85.8508960"],
+  },
+];
+
+export const empty_suggestions = [];
